% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Connects to the database and returns a database connection object}
\usage{
connect_to_database(username, password)
}
\arguments{
\item{username}{Username being used to access database}

\item{password}{Password beig used to access database}
}
\value{
Database connection object
}
\description{
Connects to the database and returns a database connection object
}
\examples{
connect_to_database("MyUserName","MyPassword")
}
