% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_das1_datawsu.R
\name{fetch_das1_datawsu}
\alias{fetch_das1_datawsu}
\title{Fetch Das1 WSU Data}
\usage{
fetch_das1_datawsu(con, deviceID, tripID)
}
\arguments{
\item{con}{PostgreSQL db connection object}

\item{deviceID}{optional search parameter, Device ID}

\item{tripID}{optional search parameter, trip ID}
}
\examples{
fetch_das1_datawsu(con, deviceID = 17103, tripID = 68)
}
