% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_das1_trip_summary.R
\name{fetch_das1_trip_summary}
\alias{fetch_das1_trip_summary}
\title{Fetch Trip Summaries}
\usage{
fetch_das1_trip_summary(con, deviceID, startdate)
}
\arguments{
\item{con}{PostgreSQL db connection object}

\item{deviceID}{optional search parameter, Device ID}

\item{startdate}{optional search parameter, trip date, format 'YYYY-MM-DD'}
}
\value{
Dataframe containing Das1_Trip_Summary table entries
}
\examples{
fetch_das1_trip_summary(con, deviceID = 10107, startdate = '2012-10-01')
}
